move local storage hook to hooks directory
