+++ .eslintrc.json
@@ -4,7 +4,7 @@
     "import/no-relative-parent-imports": [
       "error",
       {
-        "ignore": ["@/app", "@/components", "@/lib"]
+        "ignore": ["@/app", "@/components", "@/lib", "@/hooks"]
       }
     ]
   }
